backend/requirements.txt
fastapi
uvicorn
ultralytics
opencv-python-headless
python-multipart
httpx
python-dotenv